Fix bugs in vocab_log_probs propagation and numerical precision
- Add defensive bounds validation to offline_stream getVocabLogProbs()
  matching the online version
- Filter vocab_log_probs per-token in Whisper Convert() to stay aligned
  with tokens after timestamp token filtering
- Propagate vocab_log_probs through online transducer Convert() so the
  field is no longer dead on OnlineRecognizerResult
- Use double accumulator in Moonshine log-softmax to match Canary and
  avoid precision loss with large vocabularies
- Restore direct returns in Canary ForwardEncoder/ForwardDecoder for
  guaranteed copy elision

+++ sherpa-onnx/csrc/offline-moonshine-greedy-search-decoder.cc
@@ -65,11 +65,11 @@ OfflineMoonshineGreedySearchDecoder::Decode(Ort::Value encoder_out) {
     // Compute log-softmax once for both max selection and storage
     float max_logit = *std::max_element(p, p + vocab_size);
 
-    float sum_exp = 0.0f;
+    double sum_exp = 0.0;
     for (int32_t j = 0; j < vocab_size; ++j) {
       sum_exp += std::exp(p[j] - max_logit);
     }
-    float log_sum = max_logit + std::log(sum_exp);
+    float log_sum = max_logit + static_cast<float>(std::log(sum_exp));
 
     // Compute log-softmax for all tokens and find max in single pass
     std::vector<float> full_vocab_probs(vocab_size);
